# public/openapi.yaml
openapi: 3.0.1
info:
  title: Travel Plugin
  description: A plugin that allows the user to find the Position of Interests while using ChatGPT.
  version: 'v1'
servers:
  - url: https://nextjs-chatgpt-plugin-starter.vercel.app
paths:
  /api/get-poi:
    post:
      operationId: getPois
      summary: Get the list of position of interest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/searchPoiRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getPoiResponse'
  /api/get-experience:
    post:
      operationId: getExperience
      summary: Get the list of travel experience or activities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/searchExpRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getExpResponse'
components:
  schemas:
    searchPoiRequest:
      type: object
      required:
      - keyword
      properties:
        keyword:
          type: string
          description: The keyword for searching the position of interest.
          required: true
    searchExpRequest:
      type: object
      required:
      - keyword
      properties:
        keyword:
          type: string
          description: The keyword for searching the travel experience/activity.
          required: true
    getExpResponse:
      type: object
      required:
      - keyword
      properties:
        experiences:
          type: array
          items:
            type: object
            $ref: '#/components/schemas/expInfo'
          description: The list of experience/activity.
    expInfo:
      type: object
      properties:
        name:
          type: string
          description: Name of this experience/activity.
        price:
          type: number
          description: price of this experience/activity.
        rating:
          type: number
          description: Rating from 1 to 5.
        url:
          type: string
          description: Link to the experience/activity.
        cities:
          type: array
          items:
            type: object
            $ref: '#/components/schemas/cityInfo'
          description: The city information of the experience/activity.
    cityInfo:
      type: object
      properties:
        cityName:
          type: string
          description: Name of the city.
        countryName:
          type: string
          description: Name of the country.
    getPoiResponse:
      type: object
      required:
      - keyword
      properties:
        pois:
          type: array
          items:
            type: object
            $ref: '#/components/schemas/poiInfo'
          description: The list of Position of Interest.
    poiInfo:
      type: object
      properties:
        poiId:
          type: string
          description: Unique Id of LINE travel.
        poiURL:
          type: string
          description: The link to POI detail information.
        name:
          type: string
          description: Name of the POI.
        nameEn:
          type: string
          description: English name of the POI.
        nickname:
          type: array
          items:
            type: string
          description: The nicknames of the POI.
        latitude:
          type: number
          description: Latitude of the POI.
        longitude:
          type: number
          description: Longitude of the POI.
        score:
          type: number
          description: Matching score.
        coverPhoto:
          type: string
          description: The image url of cover photo of the POI.